+++ core/src/main/java/org/apache/metamodel/factory/DataContextFactory.java
@@ -16,17 +16,15 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.metamodel.util;
+package org.apache.metamodel.factory;
 
-import junit.framework.TestCase;
+import org.apache.metamodel.ConnectionException;
+import org.apache.metamodel.DataContext;
 
-public class UrlResourceTest extends TestCase {
+public interface DataContextFactory {
 
-    public void testGetName() throws Exception {
-        UrlResource resource = new UrlResource("http://metamodel.apache.org/robots.txt");
-        assertEquals("robots.txt", resource.getName());
-        
-        resource = new UrlResource("http://metamodel.apache.org/");
-        assertEquals("http://metamodel.apache.org/", resource.getName());
-    }
+    public boolean accepts(DataContextProperties properties, ResourceFactoryRegistry resourceFactoryRegistry);
+
+    public DataContext create(DataContextProperties properties, ResourceFactoryRegistry resourceFactoryRegistry)
+            throws UnsupportedDataContextPropertiesException, ConnectionException;
 }