+++ core/src/main/java/org/apache/metamodel/factory/FileResourceFactory.java
@@ -16,17 +16,23 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.metamodel.util;
+package org.apache.metamodel.factory;
 
-import junit.framework.TestCase;
+import org.apache.metamodel.util.FileResource;
+import org.apache.metamodel.util.Resource;
 
-public class UrlResourceTest extends TestCase {
+public class FileResourceFactory implements ResourceFactory {
 
-    public void testGetName() throws Exception {
-        UrlResource resource = new UrlResource("http://metamodel.apache.org/robots.txt");
-        assertEquals("robots.txt", resource.getName());
-        
-        resource = new UrlResource("http://metamodel.apache.org/");
-        assertEquals("http://metamodel.apache.org/", resource.getName());
+    @Override
+    public boolean accepts(ResourceProperties properties) {
+        final String scheme = properties.getUri().getScheme();
+        return scheme == null || "file".equals(scheme);
+    }
+
+    @Override
+    public Resource create(ResourceProperties properties) throws UnsupportedResourcePropertiesException {
+        assert accepts(properties);
+        final String path = properties.getUri().getPath();
+        return new FileResource(path);
     }
 }