+++ core/src/main/java/org/apache/metamodel/factory/ResourceFactoryRegistry.java
@@ -16,17 +16,19 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.metamodel.util;
+package org.apache.metamodel.factory;
 
-import junit.framework.TestCase;
+import java.util.Collection;
 
-public class UrlResourceTest extends TestCase {
+import org.apache.metamodel.util.Resource;
 
-    public void testGetName() throws Exception {
-        UrlResource resource = new UrlResource("http://metamodel.apache.org/robots.txt");
-        assertEquals("robots.txt", resource.getName());
-        
-        resource = new UrlResource("http://metamodel.apache.org/");
-        assertEquals("http://metamodel.apache.org/", resource.getName());
-    }
+public interface ResourceFactoryRegistry {
+
+    public void addFactory(ResourceFactory factory);
+
+    public void clearFactories();
+
+    public Collection<ResourceFactory> getFactories();
+
+    public Resource createResource(ResourceProperties properties) throws UnsupportedResourcePropertiesException;
 }